# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sales Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed, cancelled]
        - in: query
          name: customerId
          schema:
            type: integer
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "200":
          description: Updated sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /sales/{id}/invoice:
    get:
      operationId: getInvoiceData
      tags: [sales]
      summary: Get printable invoice data for a sale
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /users/{id}/role:
    patch:
      operationId: updateMemberRole
      tags: [users]
      summary: Update a team member's role (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
components:
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: ["string", "null"]
        productId:
          type: ["integer", "null"]
        productName:
          type: string
        category:
          type: string
        price:
          type: number
        quantity:
          type: integer
        subtotal:
          type: number
        discountAmount:
          type: number
        gstAmount:
          type: number
        total:
          type: number
        paymentMethod:
          type: ["string", "null"]
          enum: [cash, upi, card, bank_transfer, null]
        status:
          type: string
          enum: [pending, completed, cancelled]
        saleDate:
          type: string
          format: date-time
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        createdBySalesOsId:
          type: ["string", "null"]
        createdByName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - productName
        - category
        - price
        - quantity
        - subtotal
        - discountAmount
        - gstAmount
        - total
        - status
        - saleDate
        - createdAt
        - updatedAt
    SaleInput:
      type: object
      properties:
        productId:
          type: ["integer", "null"]
        productName:
          type: string
          minLength: 1
        category:
          type: string
          minLength: 1
        price:
          type: number
          minimum: 0
        quantity:
          type: integer
          minimum: 1
        discountAmount:
          type: number
          minimum: 0
        gstPercent:
          type: number
          minimum: 0
        paymentMethod:
          type: ["string", "null"]
          enum: [cash, upi, card, bank_transfer, null]
        status:
          type: string
          enum: [pending, completed, cancelled]
        saleDate:
          type: string
          format: date-time
        customerId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - productName
        - category
        - price
        - quantity
        - status
        - saleDate
...
    UserRole:
      type: string
      enum: [admin, manager, staff]
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        salesOsUserId:
          type: string
        email:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/UserRole"
        createdAt:
          type: string
          format: date-time
      required: [id, salesOsUserId, role, createdAt]
    RoleUpdate:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/UserRole"
      required: [role]
...
    InvoiceLine:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        amount:
          type: number
      required: [description, quantity, unitPrice, amount]
    Invoice:
      type: object
      properties:
        invoiceNumber:
          type: string
        issueDate:
          type: string
          format: date-time
        status:
          type: string
        paymentMethod:
          type: ["string", "null"]
        sellerName:
          type: string
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLine"
        subtotal:
          type: number
        discountAmount:
          type: number
        gstAmount:
          type: number
        total:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - invoiceNumber
        - issueDate
        - status
        - sellerName
        - lines
        - subtotal
        - discountAmount
        - gstAmount
        - total
...