# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sales Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface TeamMember {
  id: number;
  salesOsUserId: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  role: UserRole;
  createdAt: Date;
}